requirements.txt
django==2.2
nameko
djangorestframework
bootstrap_admin
psycopg2
